requirements.txt
numpy>=1.26,<3
pandas>=2.2,<3
Pillow>=10.0,<12
PyYAML>=6.0,<7
scikit-learn>=1.4,<2
torch>=2.2,<3
torchvision>=0.17,<1
tqdm>=4.66,<5
